Fixed Ghost buttonName issue
Fixed by figuring out I needed to set buttonName as a childObject to the button. Found a few new issues to work on, need to set up SetSpawnPos in the editor for droppers, need to fix smooth transition to work better when I have a lot of products on the conveyor (in the dictionary) or just make some more conveyor models to make the elbows of two connecting conveyors have walls, and finally I need to find all objects that are linked in my code and properly set them up as children and parent objects.